refactor(node): split configuration into focused private modules
Replace the monolithic config.rs with a public facade and private owners
for auth, network selection, journal settings, source overrides, resolved
settings, and runtime inputs. Move the three existing unit tests intact.

Preserve all 24 public config names and the six integration regressions
from the parent commit. Production item tokens and attributes are
unchanged apart from six narrowly scoped pub(super) visibility changes.

Static source and whitespace checks pass. Compiler-backed validation is
pending because the local environment has no Rust toolchain.
