+++ crates/node/src/config/network.rs
@@ -1,0 +1,76 @@
+//! Consensus-network selection and the drynet4 bootstrap profile.
+
+use std::str::FromStr;
+
+use anyhow::Result;
+use bitcoin_rs_primitives::Network;
+use serde::Deserialize;
+
+pub(super) const DRYNET4_CONNECT: &str = "drynet4.drivechain.dev:8533";
+pub(super) const DRYNET4_P2P_MAGIC: [u8; 4] = [0xec, 0xa5, 0xd4, 0x04];
+
+/// A built-in node network and its associated P2P bootstrap profile.
+#[derive(Clone, Copy, Debug, Deserialize, Eq, PartialEq)]
+#[serde(rename_all = "lowercase")]
+pub enum NetworkSelection {
+    /// Bitcoin mainnet.
+    Mainnet,
+    /// Legacy Bitcoin testnet.
+    Testnet3,
+    /// Bitcoin testnet4.
+    Testnet4,
+    /// Bitcoin signet.
+    Signet,
+    /// Local regression-test network.
+    Regtest,
+    /// ecash drynet4: mainnet consensus history on a distinct P2P network.
+    Drynet4,
+}
+
+impl NetworkSelection {
+    /// Parses the accepted network spellings.
+    #[must_use]
+    pub fn parse(value: &str) -> Option<Self> {
+        match value.trim().to_ascii_lowercase().as_str() {
+            "main" | "mainnet" | "bitcoin" => Some(Self::Mainnet),
+            "test" | "testnet" | "testnet3" => Some(Self::Testnet3),
+            "testnet4" => Some(Self::Testnet4),
+            "signet" => Some(Self::Signet),
+            "regtest" => Some(Self::Regtest),
+            "drynet4" => Some(Self::Drynet4),
+            _ => None,
+        }
+    }
+
+    /// Returns the consensus network selected by this profile.
+    #[must_use]
+    pub const fn consensus_network(self) -> Network {
+        match self {
+            Self::Mainnet | Self::Drynet4 => Network::Mainnet,
+            Self::Testnet3 => Network::Testnet3,
+            Self::Testnet4 => Network::Testnet4,
+            Self::Signet => Network::Signet,
+            Self::Regtest => Network::Regtest,
+        }
+    }
+}
+
+impl FromStr for NetworkSelection {
+    type Err = String;
+
+    fn from_str(value: &str) -> Result<Self, Self::Err> {
+        Self::parse(value).ok_or_else(|| format!("unknown network {value}"))
+    }
+}
+
+impl From<Network> for NetworkSelection {
+    fn from(network: Network) -> Self {
+        match network {
+            Network::Mainnet => Self::Mainnet,
+            Network::Testnet3 => Self::Testnet3,
+            Network::Testnet4 => Self::Testnet4,
+            Network::Signet => Self::Signet,
+            Network::Regtest => Self::Regtest,
+        }
+    }
+}
